quant_trading_system/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
psutil==5.9.6
PyJWT==2.8.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2

